+++ packages/pixer/lib/pixer.dart
@@ -6,3 +6,4 @@ export 'src/pixer_exception.dart';
 export 'src/filter_type.dart';
 export 'src/image_format.dart';
 export 'src/image_metadata.dart';
+export 'src/pixer_encoder.dart';